requirements.txt
pydantic
pyyaml
langchain
langchain-openai
redis
celery
pytest